requirements.txt
keyboard
pyyaml
pywin32
pyperclip
opencv-python
numpy
mss
rapidocr_onnxruntime
pytesseract
Pillow
